fix: Capture wildcard path correctly, resolve #167.
c70d2bc resolved the panic raised in the issue, but it also exposed my
ignorance with regards to route-recognizer, our new routing library.

Placing the * after a named param, as it was in the original example,
leads to the behavior mentioned in #167. route-recognizer is operating
exactly as intended. `:path*` doesn't capture additional URL segments,
where `*path` does.