compiler pragmas
Add some good practise pragmas.

Additionally, this also worksaround clippy bug allowing it to pass with the addition of crate level `#![allow(clippy::needless_lifetimes)]`. It's applied on a crate level so as to avoid this from being added repeatedly across the modules or functions - and since it's strictly a  temporary measure. 

Related: https://github.com/rustasync/tide/pull/176