Route: add serve_index()
I chose this way of doing the "index" thing to avoid having to do an expensive directory iteration find in order to know if and which file to serve for an "index".

This makes the developer have to specify it up-front, ensuring that it actually goes to what they want without any extra processing.

Refs: https://github.com/http-rs/tide/issues/604