src, utils: add TideState alias
This replaces `State: Clone + Send + Sync + 'static` uses with
`State: TideState`, a trait alias for the same thing.

The goal is to make it so that any future changes to these bounds
needn't be directly modified in every instance of the `State` generic,
both in Tide and also for external middleware.